Reorganise module structure
