+++ PyCO2SYS/api/__init__.py
@@ -2,6 +2,7 @@
 # Copyright (C) 2020  Matthew Paul Humphreys et al.  (GNU GPLv3)
 """Alternative APIs for executing the main CO2SYS function."""
 
+from . import ezio
 from ..engine import CO2SYS, _CO2SYS
