+++ PyCO2SYS/api/ezio/__init__.py
@@ -3,9 +3,9 @@
 # 
 """EZIO: Easy Input/Output of 'CO2SYS.xlsx'-style spreadsheets"""
 
-from ezio_utils import get_spreadsheet
-from ezio_utils import EZIO_calculate
-from ezio_utils import save_output
+from .ezio_utils import get_spreadsheet
+from .ezio_utils import EZIO_calculate
+from .ezio_utils import save_output
 
 
 def ezio(path,
@@ -66,5 +66,6 @@ def ezio(path,
                                k_fluoride = opt_k_fluoride,
                                total_borate = opt_total_borate)
     save_output(path, output_df) #.csv output, appends "_processed" to path.
-    return output_df
     print("Calculation completed and output .csv file created.")
+    return output_df
+
